+++ packages/container/package.json
@@ -15,6 +15,7 @@
     "license": "ISC",
     "source": "src/index.ts",
     "main": "dist/index.js",
+    "types": "dist/types.d.ts",
     "module": "dist/module.js",
     "files": [
         "dist"
@@ -31,8 +32,7 @@
         "url": "https://github.com/bring/hedwig-react-components/issues"
     },
     "dependencies": {
-        "@posten-hedwig/core": "^2.0.1",
-        "prop-types": "^15.8.1"
+        "@posten-hedwig/core": "^2.0.1"
     },
     "peerDependencies": {
         "react": ">=16.14.x"
